introduce application-level transactions for non-forking contexts
This commit adds support for per-application transaction management,
allowing non-forking HAF applications to group SQL operations into
reversible logical transactions independent of blockchain forks.

Key features:
- Added table `hafd.applications_transactions_register` for tracking
  application contexts, current transaction IDs, and registered tables.
- Implemented `hive.app_transaction_table_register()` to register tables
  for app-managed rollback and commit.
- Added functions for transaction control:
  - hive.app_transaction_start()
  - hive.app_transaction_commit()
  - hive.app_transaction_rollback()
  - hive.app_transaction_commit_on_table()
  - hive.app_transaction_rollback_on_table()
  - hive.app_transaction_unregister_context()
- Extended context cleanup to remove app-managed rollback contexts.
- Added documentation entry: *Application-level Transactions
  (Non-forking Contexts)* describing architecture and SQL workflow.

This feature allows lightweight, fork-independent rollback/commit
control for analytical or aggregation applications that operate in
non-forking mode.

+++ src/hive_fork_manager/context_rewind/back_from_fork.sql
@@ -66,7 +66,7 @@ END;
 $BODY$
 ;
 
-CREATE OR REPLACE FUNCTION hive.back_from_fork_one_table( _table_schema TEXT, _table_name TEXT, _shadow_table_name TEXT, _block_num_before_fork INT )
+CREATE OR REPLACE FUNCTION hive.back_from_fork_one_table( _table_schema TEXT, _table_name TEXT, _shadow_table_name TEXT, _block_num_before_fork BIGINT )
     RETURNS void
     LANGUAGE plpgsql
     VOLATILE
